More debugging hooks

+++ ts/firebase.test.ts
@@ -2,20 +2,22 @@ import { FirebaseSignalTransport } from './firebase';
 import { testSignalTransport } from './index.tests';
 import { createSignallingFirebaseTestApp } from './firebase.tests';
 
-describe('Firebase signalling', () => {
-    testSignalTransport({
-        setup: async () => {
-            const { app, collectionName } = await createSignallingFirebaseTestApp()
+if (process.env.RUN_FIREBASE_TESTS === 'true') {
+    describe('Firebase signalling', () => {
+        testSignalTransport({
+            setup: async () => {
+                const { app, collectionName } = await createSignallingFirebaseTestApp()
 
-            return {
-                signalTransportFactory: () => {
-                    return new FirebaseSignalTransport({ database: app.database(), collectionName })
-                },
-                cleanup: async () => {
-                    // await app.auth().signOut()
-                    await app.delete()
+                return {
+                    signalTransportFactory: () => {
+                        return new FirebaseSignalTransport({ database: app.database(), collectionName })
+                    },
+                    cleanup: async () => {
+                        // await app.auth().signOut()
+                        await app.delete()
+                    }
                 }
             }
-        }
+        })
     })
-})
+}
